refactor(agent): split delegate.rs into cohesive submodules (fixes #122)
Split src/agent/dispatcher/delegate.rs into a module directory to address
CodeScene warnings:
- Complex Method: execute_tool_calls (CC = 47)
- Low Cohesion: 5 distinct responsibilities

New structure:
src/agent/dispatcher/delegate/
├── mod.rs         # ChatDelegate struct, thin NativeLoopDelegate impl
├── llm_hooks.rs   # check_signals, before_llm_call, call_llm, handle_text_response
│                  # + compact_messages_for_retry, strip_internal_tool_call_text
└── tool_exec.rs   # Slimmed execute_tool_calls (orchestration only)
                  # + All helper methods extracted from original:
                  #   - record_redacted_tool_calls
                  #   - group_tool_calls (PreflightOutcome, ToolBatch)
                  #   - run_tool_batch_inline, run_tool_batch_parallel
                  #   - handle_rejected_tool
                  #   - process_runnable_tool (returns Option<String> for auth)
                  #   - maybe_emit_image_sentinel
                  #   - sanitize_output
                  #   - fold_into_context
                  # + execute_chat_tool_standalone
                  # + check_auth_required, parse_auth_result, ParsedAuthData

execute_tool_calls now only orchestrates:
1. Push assistant message with tool calls
2. Send "Thinking / executing N tool(s)..." status
3. Call record_redacted_tool_calls
4. Call group_tool_calls → destructure ToolBatch
5. Dispatch to inline or parallel execution
6. Post-flight loop calling handle_rejected_tool or process_runnable_tool
7. Return deferred-auth or NeedApproval outcome

Also consolidated preflight.rs, execution.rs, postflight.rs into delegate/
to eliminate module fragmentation.

All 3575 tests pass. No observable behavior changed.
